Correções no cod e implementação lógica no bar
