+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.instanews">
+
+    <uses-permission android:name="android.permission.VIBRATE" />
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -8,19 +10,19 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activity.HomeActivity"></activity>
-        <activity android:name=".activity.RecyclerViewActivity" />
-        <activity android:name=".activity.SplashActivity">
-        <activity android:name=".SignoActivity" />
-        <activity android:name=".NoticiaActivity" />
-        <activity android:name=".SplashActivity">
+        <activity android:name=".views.activity.SplashActivity"></activity>
+        <activity android:name=".views.activity.RecyclerViewActivity" />
+        <activity android:name=".views.activity.HomeActivity"/>
+        <activity android:name=".views.activity.SignoActivity" />
+        <activity android:name=".views.activity.NoticiaActivity" />
+        <activity android:name=".views.activity.LoginActivity" />
+        <activity android:name=".views.activity.CadastroActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+
         </activity>
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".CadastroActivity" />
     </application>
 
 </manifest>
