+++ app/src/main/java/com/example/instanews/fragment/PerfilFragment.java
@@ -1,4 +1,4 @@
-package com.example.instanews;
+package com.example.instanews.fragment;
 
 
 import android.os.Bundle;
@@ -9,15 +9,16 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.instanews.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class SearchFragment extends Fragment {
+public class PerfilFragment extends Fragment {
 
 
-
-    public SearchFragment() {
+    public PerfilFragment() {
         // Required empty public constructor
     }
 
@@ -26,7 +27,7 @@ public class SearchFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_search, container, false);
+        return inflater.inflate(R.layout.fragment_perfil, container, false);
     }
 
 }
