+++ app/src/main/java/com/example/instanews/fragment/SearchFragment.java
@@ -1,4 +1,4 @@
-package com.example.instanews;
+package com.example.instanews.fragment;
 
 
 import android.os.Bundle;
@@ -9,14 +9,17 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.instanews.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class PerfilFragment extends Fragment {
+public class SearchFragment extends Fragment {
 
 
-    public PerfilFragment() {
+
+    public SearchFragment() {
         // Required empty public constructor
     }
 
@@ -25,7 +28,7 @@ public class PerfilFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_perfil, container, false);
+        return inflater.inflate(R.layout.fragment_search, container, false);
     }
 
 }
