+++ app/src/main/res/layout-v16/activity_cadastro.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.CadastroActivity"
+    tools:context=".views.activity.CadastroActivity"
     android:orientation="vertical">
 
     <TextView
