+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context="com.example.instanews.activity.MainActivity">
+    tools:context=".views.activity.MainActivity">
 
     <TextView
         android:id="@+id/textViewTitulo"
