+++ app/src/main/res/layout/activity_signo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignoActivity">
+    tools:context=".views.activity.SignoActivity">
 
     <ScrollView
         android:layout_width="match_parent"
@@ -43,7 +43,7 @@
                 android:layout_marginRight="10dp"
                 android:layout_height="250dp"
                 android:scaleType="centerCrop"
-                app:srcCompat="@drawable/fotoSigno" />
+                app:srcCompat="@drawable/fotosigno" />
 
             <TextView
                 android:id="@+id/fonteNoticia1"
