+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SearchFragment">
+    tools:context=".fragment.SearchFragment">
 
     <SearchView
         android:layout_width="match_parent"
